(protocol-base): move blob to driver-utils (#15687)
This PR moves below classes to driver-utils and also updates the imports
to point to driver-utils:
1. BlobTreeEntry 
2. TreeTreeEntry 
3. AttachmentTreeEntry
