+++ .changeset/light-months-hope.md
@@ -1,0 +1,10 @@
+---
+"@fluidframework/container-runtime": minor
+"@fluidframework/register-collection": minor
+"@fluidframework/runtime-utils": minor
+"@fluidframework/driver-utils": minor
+"@fluidframework/tool-utils": minor
+"@fluidframework/matrix": minor
+---
+
+Move `BlobTreeEntry`, `TreeTreeEntry` and `AttachmentTreeEntry` classes from `@fluidframework/protocol-base` to `@fluidframework/driver-utils`.
