+++ packages/dds/matrix/src/serialization.ts
@@ -4,8 +4,8 @@
  */
 
 import { Serializable, IChannelStorageService } from "@fluidframework/datastore-definitions";
-import { BlobTreeEntry } from "@fluidframework/protocol-base";
 import { IFluidHandle } from "@fluidframework/core-interfaces";
+import { BlobTreeEntry } from "@fluidframework/driver-utils";
 import { IFluidSerializer } from "@fluidframework/shared-object-base";
 import { bufferToString } from "@fluidframework/common-utils";
