+++ packages/runtime/runtime-utils/src/summaryUtils.ts
@@ -11,7 +11,7 @@ import {
 	Uint8ArrayToString,
 	unreachableCase,
 } from "@fluidframework/common-utils";
-import { AttachmentTreeEntry, BlobTreeEntry, TreeTreeEntry } from "@fluidframework/protocol-base";
+import { AttachmentTreeEntry, BlobTreeEntry, TreeTreeEntry } from "@fluidframework/driver-utils";
 import {
 	ITree,
 	SummaryType,
