+++ server/routerlicious/packages/protocol-base/README.md
@@ -1,7 +1,11 @@
 # @fluidframework/protocol-base
 
-Shared protocol code for client and service to share. Manages the lifetime of Quorum and proposals that needs to be
-consistent across client and service. It also provides utilities for facilitating summary creation.
+## Purpose
+
+The purpose of @fluidframework/protocol-base is:
+
+1. Manage the lifetime of Quorum and proposals that needs to be consistent across client and service.
+1. Manage git implementation which takes in a summary object and returns it mode/type.
 
 ## Quorum and Proposal
