+++ server/routerlicious/packages/protocol-base/src/gitHelper.ts
@@ -2,19 +2,15 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-
 import * as git from "@fluidframework/gitresources";
 import {
 	FileMode,
-	IBlob,
-	IAttachment,
 	ISnapshotTreeEx,
-	ITree,
-	TreeEntry,
 	SummaryType,
 	SummaryObject,
 } from "@fluidframework/protocol-definitions";
 import { unreachableCase } from "@fluidframework/common-utils";
+
 /**
  * Take a summary object and returns its git mode.
  *
@@ -55,6 +51,7 @@ export function getGitType(value: SummaryObject): "blob" | "tree" {
 }
 
 /**
+ * NOTE: Renamed from `buildHierarchy` to `buildGitTreeHeirarchy`. There is usage of this function in loader and driver layer.
  * Build a tree hierarchy base on a flat tree
  *
  * @param flatTree - a flat tree
@@ -94,60 +91,4 @@ export function buildGitTreeHeirarchy(
 	}
 
 	return root;
-}
-
-/**
- * Basic implementation of a blob ITreeEntry
- */
-export class BlobTreeEntry {
-	public readonly mode = FileMode.File;
-	public readonly type = TreeEntry.Blob;
-	public readonly value: IBlob;
-
-	/**
-	 * Creates a blob ITreeEntry
-	 * @param path - path of entry
-	 * @param contents - blob contents
-	 * @param encoding - encoding of contents; defaults to utf-8
-	 */
-	constructor(
-		public readonly path: string,
-		contents: string,
-		encoding: "utf-8" | "base64" = "utf-8",
-	) {
-		this.value = { contents, encoding };
-	}
-}
-
-/**
- * Basic implementation of a tree ITreeEntry
- */
-export class TreeTreeEntry {
-	public readonly mode = FileMode.Directory;
-	public readonly type = TreeEntry.Tree;
-
-	/**
-	 * Creates a tree ITreeEntry
-	 * @param path - path of entry
-	 * @param value - subtree
-	 */
-	constructor(public readonly path: string, public readonly value: ITree) {}
-}
-
-/**
- * Basic implementation of an attachment ITreeEntry
- */
-export class AttachmentTreeEntry {
-	public readonly mode = FileMode.File;
-	public readonly type = TreeEntry.Attachment;
-	public readonly value: IAttachment;
-
-	/**
-	 * Creates an attachment ITreeEntry
-	 * @param path - path of entry
-	 * @param id - id of external blob attachment
-	 */
-	constructor(public readonly path: string, public readonly id: string) {
-		this.value = { id };
-	}
 }
