+++ server/routerlicious/packages/protocol-base/src/test/types/validateProtocolBasePrevious.generated.ts
@@ -16,50 +16,26 @@ type TypeOnly<T> = {
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_AttachmentTreeEntry": {"forwardCompat": false}
+* "RemovedClassDeclaration_AttachmentTreeEntry": {"forwardCompat": false}
 */
-declare function get_old_ClassDeclaration_AttachmentTreeEntry():
-    TypeOnly<old.AttachmentTreeEntry>;
-declare function use_current_ClassDeclaration_AttachmentTreeEntry(
-    use: TypeOnly<current.AttachmentTreeEntry>);
-use_current_ClassDeclaration_AttachmentTreeEntry(
-    get_old_ClassDeclaration_AttachmentTreeEntry());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_AttachmentTreeEntry": {"backCompat": false}
+* "RemovedClassDeclaration_AttachmentTreeEntry": {"backCompat": false}
 */
-declare function get_current_ClassDeclaration_AttachmentTreeEntry():
-    TypeOnly<current.AttachmentTreeEntry>;
-declare function use_old_ClassDeclaration_AttachmentTreeEntry(
-    use: TypeOnly<old.AttachmentTreeEntry>);
-use_old_ClassDeclaration_AttachmentTreeEntry(
-    get_current_ClassDeclaration_AttachmentTreeEntry());
 
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_BlobTreeEntry": {"forwardCompat": false}
+* "RemovedClassDeclaration_BlobTreeEntry": {"forwardCompat": false}
 */
-declare function get_old_ClassDeclaration_BlobTreeEntry():
-    TypeOnly<old.BlobTreeEntry>;
-declare function use_current_ClassDeclaration_BlobTreeEntry(
-    use: TypeOnly<current.BlobTreeEntry>);
-use_current_ClassDeclaration_BlobTreeEntry(
-    get_old_ClassDeclaration_BlobTreeEntry());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_BlobTreeEntry": {"backCompat": false}
+* "RemovedClassDeclaration_BlobTreeEntry": {"backCompat": false}
 */
-declare function get_current_ClassDeclaration_BlobTreeEntry():
-    TypeOnly<current.BlobTreeEntry>;
-declare function use_old_ClassDeclaration_BlobTreeEntry(
-    use: TypeOnly<old.BlobTreeEntry>);
-use_old_ClassDeclaration_BlobTreeEntry(
-    get_current_ClassDeclaration_BlobTreeEntry());
 
 /*
 * Validate forward compat by using old type in place of current type
@@ -293,26 +269,14 @@ use_old_TypeAliasDeclaration_QuorumProposalsSnapshot(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_TreeTreeEntry": {"forwardCompat": false}
+* "RemovedClassDeclaration_TreeTreeEntry": {"forwardCompat": false}
 */
-declare function get_old_ClassDeclaration_TreeTreeEntry():
-    TypeOnly<old.TreeTreeEntry>;
-declare function use_current_ClassDeclaration_TreeTreeEntry(
-    use: TypeOnly<current.TreeTreeEntry>);
-use_current_ClassDeclaration_TreeTreeEntry(
-    get_old_ClassDeclaration_TreeTreeEntry());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "ClassDeclaration_TreeTreeEntry": {"backCompat": false}
+* "RemovedClassDeclaration_TreeTreeEntry": {"backCompat": false}
 */
-declare function get_current_ClassDeclaration_TreeTreeEntry():
-    TypeOnly<current.TreeTreeEntry>;
-declare function use_old_ClassDeclaration_TreeTreeEntry(
-    use: TypeOnly<old.TreeTreeEntry>);
-use_old_ClassDeclaration_TreeTreeEntry(
-    get_current_ClassDeclaration_TreeTreeEntry());
 
 /*
 * Validate forward compat by using old type in place of current type
